Added rate limiting to create and Delete, updated Read me

+++ src/main/java/com/app/controller/Controller.java
@@ -1,6 +1,7 @@
 package com.app.controller;
 
 import java.net.URI;
+import java.time.Duration;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -23,100 +24,104 @@ import com.app.repository.UrlDataRepository;
 import com.app.service.ShortenerService;
 import com.app.validation.ValidationService;
 
+import io.github.bucket4j.Bandwidth;
+import io.github.bucket4j.Bucket;
+import io.github.bucket4j.Refill;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
-
-
 
 @RestController
 public class Controller {
 
-	// The URL shortener should support 3 methods creation, deletion, and redirect for the URLs
+	// The URL shortener should support 3 methods creation, deletion, and redirect
+	// for the URLs
 	@Autowired
 	ShortenerService shortenerService;
-	
+
 	@Autowired
 	ValidationService validationService;
-	
+
 	@Autowired
 	UrlDataRepository urlRepository;
-	
-	 private static final Logger log = LoggerFactory.getLogger(Controller.class);
 
+	private final Bucket createDeleteBucket;
+
+	private static final Logger log = LoggerFactory.getLogger(Controller.class);
+	
+	 public Controller() {
+	        Bandwidth limit = Bandwidth.classic(50, Refill.greedy(50, Duration.ofMinutes(1)));
+	        this.createDeleteBucket = Bucket.builder()
+	            .addLimit(limit)
+	            .build();
+	    }
 
 	@PostMapping("/createUrl")
-	public ResponseEntity<CreateResponse> createUrl(
-			@Valid @NotBlank @RequestParam String apiDevKey,
-			@Valid @NotBlank @RequestParam String originalUrl,
-			@RequestParam(required = false)  String customAlias,
-			@RequestParam(required = false)  LocalDate expiryDate) {
+	public ResponseEntity<CreateResponse> createUrl(@Valid @NotBlank @RequestParam String apiDevKey,
+			@Valid @NotBlank @RequestParam String originalUrl, @RequestParam(required = false) String customAlias,
+			@RequestParam(required = false) LocalDate expiryDate) {
+		
+		if(createDeleteBucket.tryConsume(1)) {
+			log.info("Calling Create URL with params: " + " apiDevKey= " + " originalUrl= " + " customAlias= "
+					+ " expiryDate= ", apiDevKey, originalUrl, customAlias, expiryDate);
+			
+			validationService.validateCreateUrl(apiDevKey, originalUrl, customAlias, expiryDate);
 
-		log.info("Calling Create URL with params: "+" apiDevKey= "+" originalUrl= "+" customAlias= "+ " expiryDate= ",apiDevKey,originalUrl,customAlias,expiryDate);
-		
-		validationService.validateCreateUrl(apiDevKey,originalUrl,customAlias,expiryDate);
+			UrlData createdUrl = shortenerService.createURL(apiDevKey, originalUrl, customAlias, expiryDate);
 
-		UrlData createdUrl= shortenerService.createURL(apiDevKey, originalUrl,customAlias, expiryDate);
-		
+			CreateResponse response = CreateResponse.builder().originalUrl(originalUrl).shortUrl(createdUrl.getShortUrl())
+					.code(HttpStatus.OK).expiration(createdUrl.getExpiryDate()).response("succeeded").build();
 
-		CreateResponse response = CreateResponse.builder().
-				originalUrl(originalUrl)
-				.shortUrl(createdUrl.getShortUrl())
-				.code(HttpStatus.OK)
-				.expiration(createdUrl.getExpiryDate())
-				.response("succeeded")
-				.build();
+			log.info("Created Url Response: ", response);
+
+			return new ResponseEntity<>(response, response.getCode());
+			
+		}
 		
-		log.info("Created Url Response: ",response);
-		
-		
-		return new ResponseEntity<>(response, response.getCode());
+		 return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
 
 	}
 
 	@GetMapping("/deleteUrl")
-	public ResponseEntity<DeleteResponse> deleteUrl(
-			@Valid @NotBlank @RequestParam String apiDevKey,
+	public ResponseEntity<DeleteResponse> deleteUrl(@Valid @NotBlank @RequestParam String apiDevKey,
 			@Valid @NotBlank @RequestParam String urlKey) {
-		
-		int count= shortenerService.deleteUrl(apiDevKey, urlKey);
 
-		DeleteResponse response = generateDeleteResponse(urlKey,count);
-		
-		log.info("Deleting URL Response: ", response);
-		
-		return new ResponseEntity<>(response, response.getCode());
+		if(createDeleteBucket.tryConsume(1)) {
+			int count = shortenerService.deleteUrl(apiDevKey, urlKey);
 
+			DeleteResponse response = generateDeleteResponse(urlKey, count);
 
+			log.info("Deleting URL Response: ", response);
+
+			return new ResponseEntity<>(response, response.getCode());
+		}
+		
+		 return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
 	}
 
-	private DeleteResponse generateDeleteResponse(String urlKey,int count) {
-		
-		String result= count==0 ? "Failed to Delete":"Deleted";
-		HttpStatus code= count ==0 ? HttpStatus.BAD_REQUEST: HttpStatus.OK;
-		String status= code==HttpStatus.OK ? "success" : "fail";
-		
-		DeleteResponse response =  DeleteResponse.builder()
-				.shortUrl(urlKey)
-				.code(code)
-				.status(status)
-				.response(result)
+	private DeleteResponse generateDeleteResponse(String urlKey, int count) {
+
+		String result = count == 0 ? "Failed to Delete" : "Deleted";
+		HttpStatus code = count == 0 ? HttpStatus.BAD_REQUEST : HttpStatus.OK;
+		String status = code == HttpStatus.OK ? "success" : "fail";
+
+		DeleteResponse response = DeleteResponse.builder().shortUrl(urlKey).code(code).status(status).response(result)
 				.build();
-		
+
 		return response;
 	}
 
 	@GetMapping("/redirectUrl")
 	public ResponseEntity<Void> redirectUrl(@Valid @NotBlank @RequestParam String apiDevKey,
 			@Valid @NotBlank @RequestParam String shortUrl) {
-		
-		log.info("Redirecting URL: ",shortUrl);
 
-		validationService.validateRedirectUrl(apiDevKey,shortUrl);	
-	
+		log.info("Redirecting URL: ", shortUrl);
+
+		validationService.validateRedirectUrl(apiDevKey, shortUrl);
+
 		String redirectUrlPath = shortenerService.redirectUrl(apiDevKey, shortUrl);
-		
+
 		URI redirectUri = URI.create(redirectUrlPath);
-	   
+
 		log.info("Redirection URL Response: ", redirectUri);
 		return ResponseEntity.status(HttpStatus.FOUND).location(redirectUri).build();
 
@@ -124,14 +129,13 @@ public class Controller {
 
 	@GetMapping("/getAll")
 	public List<UrlData> GetAll() {
-		
+
 		log.info("Getting All Items");
-		
-		List<UrlData> response= urlRepository.findAll();
+
+		List<UrlData> response = urlRepository.findAll();
 		return response;
 	}
-	
-	
+
 	@GetMapping("/health")
 	public String healthCheck() {
 		log.info("Running Health Check");
